+++ tests/unit/interface/no_tango/test_cm_set_kvalue.py
@@ -1,7 +1,7 @@
 """Tests dish manager component manager setkvalue command handler"""
 
 import pytest
-from ska_control_model import CommunicationStatus, ResultCode
+from ska_control_model import ResultCode
 
 from ska_mid_dish_manager.component_managers.dish_manager_cm import DishManagerComponentManager
 
@@ -19,10 +19,6 @@ def test_set_kvalue_handler(
         the command tracker under test
     """
     # set_kvalue has no pre-condition.
-    # It only requires a connection to the spfrx device
-    rx_component_manager = component_manager.sub_component_managers["SPFRX"]
-    rx_component_manager._update_communication_state(CommunicationStatus.ESTABLISHED)
-
     status, message = component_manager.set_kvalue(5)
 
     # this is a fast command and returns immediately