+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 - Device State reports ON when device is started.
 - Periodic calls to `MonitorPing` on SPFRx is handled by component manager instead of polled command.
+- Log handling included for when the device's communication state is `DISABLED` or `NOT_ESTABLISHED`
 
 Version 7.0.1
 *************